@@ -4,10 +4,10 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.5.7",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/lab": "^5.0.0-alpha.66",
-    "@mui/material": "^5.2.8",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/lab": "^5.0.0-alpha.78",
+    "@mui/material": "^5.6.2",
     "@react-spring/web": "^9.4.2",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
@@ -17,28 +17,29 @@
     "@types/leaflet": "^1.7.9",
     "@types/lodash": "^4.14.172",
     "@types/node": "^12.20.19",
-    "@types/react": "^17.0.16",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
     "autosuggest-highlight": "^3.2.0",
+    "buffer": "^6.0.3",
     "date-fns": "^2.28.0",
     "graphql": "^16.2.0",
     "i18next": "^21.6.14",
     "i18next-browser-languagedetector": "^6.1.4",
     "i18next-http-backend": "^1.4.0",
-    "leaflet": "npm:@findev/leaflet@^1.8.6",
-    "leaflet.vectorgrid": "^1.3.0",
+    "leaflet": "^1.8.0",
     "lodash": "^4.17.21",
     "mdi-material-ui": "^7.1.0",
-    "mqtt": "^4.3.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.16.2",
-    "react-leaflet": "^3.2.5",
-    "react-move": "^6.5.0",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^4.0.3",
-    "react-svg": "^14.1.14",
+    "mqtt": "^4.3.7",
+    "process": "^0.11.10",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.16.7",
+    "react-leaflet": "^4.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-svg": "^14.1.18",
     "typescript": "^4.5.5",
+    "url": "^0.11.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
