@@ -1,8 +1,10 @@
+import './utils/mqttPatch';
+
 import React, { Suspense } from 'react';
 
 import { ApolloProvider } from '@apollo/client';
 import { CircularProgress } from '@mui/material';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import './index.css';
 import App from './App';
@@ -10,7 +12,9 @@ import { client } from './graphql/client';
 import './i18n';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(
   <React.StrictMode>
     <ApolloProvider client={client}>
       <Suspense
@@ -23,8 +27,7 @@ ReactDOM.render(
         <App />
       </Suspense>
     </ApolloProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
